Separated security related API into the ripple-security module (should be no obligation to use it)
