+++ ripple-security/src/main/java/org/rippleosi/api/login/rest/LoginController.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.rippleosi.security.login.rest;
+package org.rippleosi.api.login.rest;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
