+++ ripple-security/src/main/java/org/rippleosi/config/rest/SecurityRestConfig.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2016 Ripple OSI
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.rippleosi.config.rest;
+
+import org.pac4j.core.config.Config;
+import org.pac4j.springframework.web.RequiresAuthenticationInterceptor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+/**
+ */
+@Configuration
+@EnableWebMvc
+@ComponentScan("org.rippleosi")
+public class SecurityRestConfig extends WebMvcConfigurerAdapter {
+
+    @Autowired
+    private Config config;
+
+    @Bean
+    public RequiresAuthenticationInterceptor getAuthenticationHandler() {
+        return new RequiresAuthenticationInterceptor(config, "OidcClient");
+    }
+}
