+++ ripple-security/src/main/java/org/rippleosi/config/security/SecurityConfig.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.rippleosi.config;
+package org.rippleosi.config.security;
 
 import org.pac4j.core.config.Config;
 import org.rippleosi.common.interceptor.authentication.RequiresAuthenticationInterceptor;
